pyproject.toml
[project]
name = "learner-portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "kuzu>=0.11.3",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
    "ty>=0.0.21",
]
